+++ frame/bridge/relay-authorities/Cargo.toml
@@ -13,6 +13,7 @@ version     = "2.8.15"
 # crates.io
 codec      = { package = "parity-scale-codec", version = "2.3", default-features = false, features = ["derive"] }
 scale-info = { version = "1.0", default-features = false, features = ["derive"] }
+serde      = { version = "1.0", optional = true }
 # darwinia-network
 darwinia-header-mmr       = { default-features = false, path = "../../../frame/header-mmr" }
 darwinia-relay-primitives = { default-features = false, path = "../../../primitives/relay" }
@@ -20,6 +21,7 @@ darwinia-support          = { default-features = false, path = "../../support" }
 # paritytech
 frame-support = { default-features = false, git = "https://github.com/darwinia-network/substrate", branch = "darwinia-v0.12.3" }
 frame-system  = { default-features = false, git = "https://github.com/darwinia-network/substrate", branch = "darwinia-v0.12.3" }
+sp-io         = { default-features = false, git = "https://github.com/darwinia-network/substrate", branch = "darwinia-v0.12.3" }
 sp-runtime    = { default-features = false, git = "https://github.com/darwinia-network/substrate", branch = "darwinia-v0.12.3" }
 sp-std        = { default-features = false, git = "https://github.com/darwinia-network/substrate", branch = "darwinia-v0.12.3" }
 
@@ -42,6 +44,7 @@ std = [
 	# crates.io
 	"codec/std",
 	"scale-info/std",
+	"serde",
 	# darwinia-network
 	"darwinia-header-mmr/std",
 	"darwinia-relay-primitives/std",
@@ -51,4 +54,5 @@ std = [
 	"frame-system/std",
 	"sp-runtime/std",
 	"sp-std/std",
+	"sp-io/std",
 ]
