+++ frame/wormhole/backing/ethereum/src/tests.rs
@@ -27,7 +27,6 @@ use crate::{
 	pallet::*,
 };
 use darwinia_bridge_ethereum::EthereumRelayHeaderParcel;
-use darwinia_relay_primitives::Sign;
 use darwinia_staking::{RewardDestination, StakingBalance, StakingLedger, TimeDepositItem};
 use ethereum_primitives::receipt::EthereumReceiptProof;
 
@@ -723,15 +722,6 @@ fn lock_limit_should_work() {
 			<Error<Test>>::KtonLockLim
 		);
 	});
-}
-
-#[test]
-fn verify_signature_should_work() {
-	assert!(EthereumBacking::verify_signature(
-		&array_bytes::hex2array_unchecked("0x0806e7b411a8808c1384bd8abe3b506403981d3ece6b16cd29d3f2789eea1ab61635b3b971bf5584bdc70c42b2a4a2659b354dfc542943c030630168825976491c"),
-		&array_bytes::hex2array_unchecked("0x71e2f60faf6c7264cca14fb1a01260a787b4d18039cd8cd680aaff1e118c711d"),
-	 	&array_bytes::hex_into_unchecked("0x6aA70f55E5D770898Dd45aa1b7078b8A80AAbD6C")
-	));
 }
 
 #[test]
