+++ primitives/relay/Cargo.toml
@@ -13,7 +13,6 @@ version     = "2.8.15"
 # crates.io
 codec      = { package = "parity-scale-codec", version = "2.3", default-features = false }
 scale-info = { version = "1.0", default-features = false, features = ["derive"] }
-serde      = { version = "1.0", optional = true }
 # paritytech
 frame-support = { default-features = false, git = "https://github.com/darwinia-network/substrate", branch = "darwinia-v0.12.3" }
 sp-runtime    = { default-features = false, git = "https://github.com/darwinia-network/substrate", branch = "darwinia-v0.12.3" }
@@ -26,7 +25,6 @@ std = [
 	# crates.io
 	"codec/std",
 	"scale-info/std",
-	"serde",
 	# paritytech
 	"frame-support/std",
 	"sp-runtime/std",
